vo 오타 수정
